// types/types.ts
export type Pagination<T> = {
  count: number;
  next: string | null;
  previous: string | null;
  results: T[]
}

export interface ApiResponse extends Pagination<any> {
  error?: any;
}

/* TIKCET */
export interface Ticket {
  ticket_id: string;
  title: string;
  description: string;
  status: string;
  client_phonenumber: number | TicketClientPhoneNumber | null;
  caller_email: string | TicketClientPhoneNumber | null;
  caller_phonenumber: string | null;
  email_id: string | null;
  priority: string;
  flag: string;
  cat: number | { id: number; title: string; };
  assigned_agent: { id: string; first_name: string; last_name: string; email: string, phone_number: string, profile_pic: string };
  escalation: null | EscalationForm;
  timestamp: string;
  last_updated: string | null;
}

export interface EscalationForm {
  id: number;
  escalated_at: Date;
  raised_by: number | { id: number; first_name: string; last_name: string; email: string; phone_number: string };
  reason: string;
}

export interface TicketClientPhoneNumber { first_name: string, last_name: string, id: number, phone_number: string }

export interface TicketComments {
  id: string;
  title: string;
  comments: Comment[];
}

export interface Comment {
  id: number;
  user: User;
  ticket: string;
  // title: string;
  body: string;
  timestamp: Date;
  last_updated: null | Date;
}

export interface User {
  first_name: string;
  last_name: string;
  phone_number: string | null;
}